Adding JavaFX Dependencies
