+++ src/client/gui/src/main/java/com/kdyncs/dragonsong/client/common/gui/Launcher.java
@@ -1,0 +1,26 @@
+package com.kdyncs.dragonsong.client.common.gui;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Launcher {
+
+    // Logging
+    private static final Logger log = LogManager.getLogger();
+
+    public static void main(String[] args) {
+
+        // Create a Spring Application
+        SpringApplication application = new SpringApplication(Launcher.class);
+
+        // This application does not need a web server
+        application.setWebApplicationType(WebApplicationType.NONE);
+
+        // Run Application
+        application.run(args);
+    }
+}
